requirements.txt
phidata==2.7.2
streamlit==1.41.1
